Only call the column's case_sensitive? if it is determined by the collation.
When determining whether a uniqueness validation should be case insensitive,
we only want to consider the column's own reported case-sensitivity if it
is a true property of the column, i.e., due to its collation/character set.

The schema_plus_columns gem monkey patches in its own version of
case_sensitive?, for PostgreSQL, that determines this based upon the
properties of ANY indexes in which the column participates. That
method does NOT correctly take into account the index scope.  Since in
PostgreSQL it is possible to have a column participate in two indexes,
one being case-sensitive and another being case-insensitive, it is
important to always consider the scope.  This does not matter with MySQL.