Fix the filtering of attributes validated with SchemaValidations::Validators::NotNilValidator
This appears to be a special case where the `name` parameter of
`#_filter_validation` is set to `SchemaValidations::Validators::NotNilValidator`
and the actual name is found in the `:attribute` option in `opts`.
Filtering-by-name using the `only`, `except`, etc. configurations see
the wrong `name` value in these cases and will not be filtered
correctly as a result.